# fs: Enable returning ETag along with ListObjects()