--- a/cmd/fs-v1-metadata.go
+++ b/cmd/fs-v1-metadata.go
@@ -27,6 +27,7 @@ import (
 
 	"github.com/minio/minio/pkg/lock"
 	"github.com/minio/minio/pkg/mimedb"
+	"github.com/tidwall/gjson"
 )
 
 const (
@@ -66,7 +67,7 @@ func (m fsMetaV1) ToObjectInfo(bucket, object string, fi os.FileInfo) ObjectInfo
 		Name:   object,
 	}
 
-	// We set file into only if its valid.
+	// We set file info only if its valid.
 	objInfo.ModTime = timeSentinel
 	if fi != nil {
 		objInfo.ModTime = fi.ModTime()
@@ -144,29 +145,76 @@ func (m *fsMetaV1) WriteTo(lk *lock.LockedFile) (n int64, err error) {
 	return int64(len(metadataBytes)), nil
 }
 
+func parseFSVersion(fsMetaBuf []byte) string {
+	return gjson.GetBytes(fsMetaBuf, "version").String()
+}
+
+func parseFSFormat(fsMetaBuf []byte) string {
+	return gjson.GetBytes(fsMetaBuf, "format").String()
+}
+
+func parseFSRelease(fsMetaBuf []byte) string {
+	return gjson.GetBytes(fsMetaBuf, "minio.release").String()
+}
+
+func parseFSMetaMap(fsMetaBuf []byte) map[string]string {
+	// Get xlMetaV1.Meta map.
+	metaMapResult := gjson.GetBytes(fsMetaBuf, "meta").Map()
+	metaMap := make(map[string]string)
+	for key, valResult := range metaMapResult {
+		metaMap[key] = valResult.String()
+	}
+	return metaMap
+}
+
+func parseFSParts(fsMetaBuf []byte) []objectPartInfo {
+	// Parse the FS Parts.
+	partsResult := gjson.GetBytes(fsMetaBuf, "parts").Array()
+	partInfo := make([]objectPartInfo, len(partsResult))
+	for i, p := range partsResult {
+		info := objectPartInfo{}
+		info.Number = int(p.Get("number").Int())
+		info.Name = p.Get("name").String()
+		info.ETag = p.Get("etag").String()
+		info.Size = p.Get("size").Int()
+		partInfo[i] = info
+	}
+	return partInfo
+}
+
 func (m *fsMetaV1) ReadFrom(lk *lock.LockedFile) (n int64, err error) {
-	var metadataBytes []byte
+	var fsMetaBuf []byte
 	fi, err := lk.Stat()
 	if err != nil {
 		return 0, traceError(err)
 	}
 
-	metadataBytes, err = ioutil.ReadAll(io.NewSectionReader(lk, 0, fi.Size()))
+	fsMetaBuf, err = ioutil.ReadAll(io.NewSectionReader(lk, 0, fi.Size()))
 	if err != nil {
 		return 0, traceError(err)
 	}
 
-	if len(metadataBytes) == 0 {
+	if len(fsMetaBuf) == 0 {
 		return 0, traceError(io.EOF)
 	}
 
-	// Decode `fs.json` into fsMeta structure.
-	if err = json.Unmarshal(metadataBytes, m); err != nil {
-		return 0, traceError(err)
-	}
+	// obtain version.
+	m.Version = parseFSVersion(fsMetaBuf)
+
+	// obtain format.
+	m.Format = parseFSFormat(fsMetaBuf)
+
+	// obtain metadata.
+	m.Meta = parseFSMetaMap(fsMetaBuf)
+
+	// obtain parts info list.
+	m.Parts = parseFSParts(fsMetaBuf)
+
+	// obtain minio release date.
+	m.Minio.Release = parseFSRelease(fsMetaBuf)
 
 	// Success.
-	return int64(len(metadataBytes)), nil
+	return int64(len(fsMetaBuf)), nil
 }
 
 // FS metadata constants.