--- a/cmd/test-utils_test.go
+++ b/cmd/test-utils_test.go
@@ -67,6 +67,9 @@ func init() {
 
 	// Set system resources to maximum.
 	setMaxResources()
+
+	// Quiet logging.
+	log.logger.Hooks = nil
 }
 
 func prepareFS() (ObjectLayer, string, error) {