@krishnasrinivas 2017-04-04
* what is the performance difference between gjson and json.Umnarshal?
* what is the performance difference if the following is replaced:
```
	fsMetaBuf, err := ioutil.ReadAll(rlk.LockedFile)
 +	if err != nil {
 +		// `fs.json` can be empty due to previously failed
 +		// PutObject() transaction, if we arrive at such
 +		// a situation we just ignore and continue.
 +		if errorCause(err) != io.EOF {
 +			return "", toObjectErr(err, bucket, entry)
 +		}
 +	}
 +
 +	fsMetaMap := parseFSMetaMap(fsMetaBuf)
```
by
fsMetaV1.ReadFrom() i.e will the extra work of doing parseFSVersion parseFSFormat parseFSParts parseFSRelease instead of just parseFSParts result in too much of performance hit?