@harshavardhana 2017-04-04
> what is the performance difference between gjson and json.Umnarshal?

json.Unmarshal is orders of magnitude slower than gjson and even uses lot of memory,  this was already established when we moved our xl.json reading to gjson, we never did for `fs` because we had `fs.json` optional and it was kind of not necessary. But it is the right time to move as well since we heavily depend on it and it is not optional anymore. 

It is quite simple to see this with simple bench-marking. 

With json.Unmarshal
```
$ cat old.txt 
BenchmarkParseAll10-4      	   50000	     35084 ns/op	    3064 B/op	      68 allocs/op
BenchmarkParseAll100-4     	    5000	    235998 ns/op	   27368 B/op	     440 allocs/op
BenchmarkParseAll1000-4    	    1000	   2377073 ns/op	  227696 B/op	    4050 allocs/op
BenchmarkParseAll5000-4    	     100	  12691447 ns/op	 1130992 B/op	   20058 allocs/op
BenchmarkParseAll10000-4   	      50	  24660559 ns/op	 2434096 B/op	   40062 allocs/op
PASS
ok  	_/tmp/testing	8.428s
```

With gjson
```
$ cat new.txt 
BenchmarkParseAll10-4      	  100000	     14288 ns/op	    8096 B/op	      16 allocs/op
BenchmarkParseAll100-4     	   10000	    114355 ns/op	   65984 B/op	      19 allocs/op
BenchmarkParseAll1000-4    	    2000	   1056521 ns/op	  639936 B/op	      22 allocs/op
BenchmarkParseAll5000-4    	     300	   5869262 ns/op	 3769280 B/op	      29 allocs/op
BenchmarkParseAll10000-4   	     100	  12006275 ns/op	 7717824 B/op	      32 allocs/op
PASS
ok  	_/tmp/testing	8.550s
```

```
benchcmp old.txt new.txt 
benchmark                    old ns/op     new ns/op     delta
BenchmarkParseAll10-4        35084         14288         -59.27%
BenchmarkParseAll100-4       235998        114355        -51.54%
BenchmarkParseAll1000-4      2377073       1056521       -55.55%
BenchmarkParseAll5000-4      12691447      5869262       -53.75%
BenchmarkParseAll10000-4     24660559      12006275      -51.31%

benchmark                    old allocs     new allocs     delta
BenchmarkParseAll10-4        68             16             -76.47%
BenchmarkParseAll100-4       440            19             -95.68%
BenchmarkParseAll1000-4      4050           22             -99.46%
BenchmarkParseAll5000-4      20058          29             -99.86%
BenchmarkParseAll10000-4     40062          32             -99.92%

benchmark                    old bytes     new bytes     delta
BenchmarkParseAll10-4        3064          8096          +164.23%
BenchmarkParseAll100-4       27368         65984         +141.10%
BenchmarkParseAll1000-4      227696        639936        +181.05%
BenchmarkParseAll5000-4      1130992       3769280       +233.27%
BenchmarkParseAll10000-4     2434096       7717824       +217.07%
```