@harshavardhana 2017-04-04
Especially when it comes to parsing we don't need to parse the entire struct to get to the final value, so individually for example multipart can just use parseFSParts(). Also we already well understand its behavior due to our extensive usage in `XL` . 

To go back and see what was fixed is this https://github.com/minio/minio/issues/2208