@harshavardhana 2017-04-04
More benchmarks with some parallel testing as well. old is `json.Unmarshal` and new is `gjson` 

```
$ benchcmp old.txt new.txt 
benchmark                            old ns/op     new ns/op     delta
BenchmarkParseAllParallel100-4       110276        53857         -51.16%
BenchmarkParseAllParallel500-4       548201        281849        -48.59%
BenchmarkParseAllParallel1000-4      1137198       571913        -49.71%
BenchmarkParseAllParallel5000-4      5839622       2980803       -48.96%
BenchmarkParseAllParallel10000-4     12387503      6301408       -49.13%

benchmark                            old allocs     new allocs     delta
BenchmarkParseAllParallel100-4       440            19             -95.68%
BenchmarkParseAllParallel500-4       2048           21             -98.97%
BenchmarkParseAllParallel1000-4      4050           22             -99.46%
BenchmarkParseAllParallel5000-4      20058          29             -99.86%
BenchmarkParseAllParallel10000-4     40062          32             -99.92%
```