@harshavardhana 2017-04-04
For sample benchmarking code is here - https://gist.github.com/harshavardhana/a7c1773c00ab81e66b16403f12f9015b (credits to @hackintoshrao on previous implementation)